[feat] Replace Parallel Arrays with Array of Structs
Replaced the parallel arrays with an array of structs with the same relevent info. Cashier was also modified to work using actual books. editBook and deleteBook still need to be finished, and the fact that the array of structs is global is not the best solution and should be changed to a local array at some point.